requirements.txt
# Additional Python dependencies for ISIC
# (beyond what's included in the official Odoo image)

# Redis session store
redis>=5.0
# LDAP authentication
python-ldap>=3.4
# OCA test helper (FakeModelLoader for tier validation tests)
odoo-test-helper>=2.0
# GED full-text extraction
pypdf>=4.0
python-docx>=1.0
openpyxl>=3.1
